# swagger-output.json
{
	"swagger": "2.0",
	"info": {
		"title": "Contacts API",
		"description": "Contacts API Documentation",
		"version": "1.0.1"
	},
	"host": "localhost:5500",
	"basePath": "/",
	"schemes": ["https"],
	"paths": {
		"/": {
			"get": {
				"description": "",
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/dg": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": ""
					}
				}
			},
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"diagnosis": {
									"example": "any"
								},
								"type": {
									"example": "any"
								},
								"overview": {
									"example": "any"
								},
								"symptoms": {
									"example": "any"
								},
								"treatments": {
									"example": "any"
								},
								"specialists": {
									"example": "any"
								},
								"contagiousMethod": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"201": {
						"description": "Created"
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/dg/update/{id}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"404": {
						"description": "Not Found"
					}
				}
			},
			"put": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"diagnosis": {
									"example": "any"
								},
								"type": {
									"example": "any"
								},
								"overview": {
									"example": "any"
								},
								"symptoms": {
									"example": "any"
								},
								"treatments": {
									"example": "any"
								},
								"specialists": {
									"example": "any"
								},
								"contagiousMethod": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/dg/delete/{id}": {
			"delete": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/getalldg": {
			"get": {
				"description": "",
				"responses": {
					"200": {
						"description": "OK"
					}
				}
			}
		},
		"/dg/{id}": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/success": {
			"get": {
				"description": "",
				"responses": {
					"default": {
						"description": ""
					}
				}
			}
		}
	}
}